backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
PyPDF2==3.0.1
pymupdf==1.24.0
openai==1.14.0
anthropic==0.7.0
sqlalchemy==2.0.23
mysql-connector-python==8.2.0
requests==2.31.0
httpx==0.24.1
